fix(pydantic-core-cpp): resolve EitherInt(long long) overload ambiguity on Linux
validate_int stored strtoll's result in a `long long`, which is ambiguous
against EitherInt(int64_t)/EitherInt(uint64_t) when int64_t is `long` (Linux
gcc); on macOS int64_t is `long long` so it matched exactly and built. Declare
the parsed value as int64_t (matching the existing `int64_t i = std::stoll(...)`
sites) so the constructor call is an exact match on every platform.

+++ pydantic-core-cpp/src/input/python_input.cpp
@@ -436,7 +436,7 @@ ValResult<ValMatch<EitherInt>> PythonInput::validate_int(bool strict) const {
                 }
                 errno = 0;
                 char* endp = nullptr;
-                long long v = std::strtoll(cleaned.c_str(), &endp, 10);
+                int64_t v = std::strtoll(cleaned.c_str(), &endp, 10);
                 if (errno == ERANGE && v > 0) {
                     // Beyond i64 — fall back to Python's arbitrary-precision int
                     // (Rust EitherInt::BigInt via jiter)
